grader_backend/requirements.txt
fastapi
uvicorn
python-dotenv
pymupdf
python-docx
openai
